refactor(collisions): move CollidingWith component to base collisions folder
